refactoring tests pass
