+++ DB.md
@@ -23,6 +23,7 @@ DROP DATABASE template1;
 CREATE DATABASE template1 WITH TEMPLATE = template0 ENCODING = 'UTF8';
 UPDATE pg_database SET datistemplate = TRUE WHERE datname = 'template1';
 \c template1
+
 VACUUM FREEZE;
 ```
 
@@ -46,10 +47,18 @@ and then
 psql -d inspection.canada.ca
 ```
 
+
+## postgresql extensions
+
+```
+pip install pgxnclient
+pgxn install vector
+```
+
 ## loading dump
 
 ```
-psql -v ON_ERROR_STOP=1 --single-transaction -d inspection.canada.ca < inspection.canada.ca.2023-06-09.pg_dump
+./load-db.sh dumps/inspection.canada.ca.2023-06-09.pg_dump
 ```
 
 ## workaround for psycopg2 not finding the socket file
@@ -68,12 +77,6 @@ sudo ln -s /tmp/.s.PGSQL.5432 /var/run/postgresql/.s.PGSQL.5432
 ```
 
 
-## postgresql extensions
-
-```
-pip install pgxnclient
-pgxn install vector
-```
 
 see extensions available: https://pgxn.org/
