+++ requirements.txt
@@ -15,6 +15,7 @@ cryptography==40.0.2
 cssselect==1.2.0
 filelock==3.12.0
 Flask==2.3.2
+flask-unittest==0.1.3
 frozenlist==1.3.3
 hyperlink==21.0.0
 idna==3.4
