+++ scrapy.cfg
@@ -4,7 +4,7 @@
 # https://scrapyd.readthedocs.io/en/latest/deploy.html
 
 [settings]
-default = louis.settings
+default = louis.crawler.settings
 
 [deploy]
 #url = http://localhost:6800/
