+++ tests/test_smartsearch.py
@@ -3,7 +3,7 @@ import flask_unittest
 
 class TestSmartSearch(flask_unittest.AppTestCase):
     def create_app(self):
-        from smartsearch import app
+        from louis.api.smartsearch import app
         return app
 
     def test_basic_query(self, app):
